requirements/base.txt
Django==1.5.1
django-braces==1.0.0
django-model-utils==1.4.0
django-auth-ldap==1.1.4
django-dajax==0.9.2
django-dajaxice==0.5.5
django-datatables-view==1.6
logutils==0.3.3
requests==1.2.3
South==0.7.6
reportlab==2.7
PIL==1.1.7
Pillow==2.0.0
python-dateutil==2.1
colorama==0.2.5
unipath==1.0
raven==3.4.0

# Manual Install
python-ldap==2.4.13
mysql-python==1.2.4
cx_Oracle==5.1.2